+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Changelog
 ## [Unreleased]
+### Added
+- `_is_analyzed_manually` feature in `BoostingFeaturizer` class, by @HardNorth
 ### Fixed
 - Text similarity calculation for cases when a text contains only stop words, by @HardNorth
 ### Changed